Move app core into the root
