+++ src/domain.ts
